+++ docs/fr_FR/Changelog.md
@@ -4,6 +4,7 @@
 -   Améliorations: Modèles 'CLA60 TW OSRAM' & 'ClassicB40TW-LIGHTIFY'.
 -   Améliorations: check_json: Vérification valeurs de 'Polling'.
 -   Améliorations: Assistant modèle pour device 'lightning'.
+-   Améliorations: Interne Abeille.class (publishMosquitto => msgToCmd).
 
 ## 250318-BETA-1